Don't strip #.* from ends of lines
Revert e99ab2e9d07573e24e6d106877ff35eab46cdab1, which was intended to
strip comments from lines fed to pip, but had the unintended side
effect of stripping egg specifiers in URLs. The easiest way to address
this is to simply not put trailing comments in requirements files, so
we'er going to solve the problem that way instead.