use current_data instead of last_sale_price
It seems more clear to get price values from `self.trading_client.current_data[sid].price` than from `self.portfolio.positions[sid].last_sale_price`. The two values are the same, so this is just a readability change, but it is also the same behavior as in `self.order_value()` and it's good to have them all be the same.